BackEnd/requirement.txt
Django
djangorestframework
drf-spectacular
django-cors-headers